Rename packages to follow best naming practices
