+++ src/main/java/commands/ExitCommand.java
@@ -1,7 +1,7 @@
 package commands;
 
-import helpers.TaskList;
-import helpers.Ui;
+import utils.TaskList;
+import utils.Ui;
 
 /**
  * The ExitCommand class represents a command to exit the application.
