+++ src/main/java/utils/Ui.java
@@ -1,4 +1,4 @@
-package helpers;
+package utils;
 
 import tasks.Task;
 import java.util.Scanner;
