+++ src/test/java/utils/TaskListTest.java
@@ -1,4 +1,4 @@
-package helpers;
+package utils;
 
 import mocks.MockTodo;
 import org.junit.jupiter.api.Test;
